+++ packages/snap-client/src/Client/Client.test.ts
@@ -3,6 +3,7 @@ import { Client } from './Client';
 import type { ClientConfig } from '../types';
 import { MockData } from '@searchspring/snap-shared';
 import { AppMode, version } from '@searchspring/snap-toolbox';
+import { NO_BEACON_PARAM } from './transforms';
 
 const mockData = new MockData();
 
@@ -366,10 +367,11 @@ describe('Snap Client', () => {
 					],
 					test: true,
 					siteId: '8uyt2m',
+					[NO_BEACON_PARAM]: true,
 				},
 			};
 
-			const recommendCacheKey = '{"profiles":[{"tag":"dress"}],"siteId":"8uyt2m","test":true}';
+			const recommendCacheKey = `{"profiles":[{"tag":"dress"}],"siteId":"8uyt2m","${NO_BEACON_PARAM}":true,"test":true}`;
 
 			expect(recommendRequesterSpy).toHaveBeenCalledTimes(2);
 			expect(recommendRequesterSpy.mock.calls).toEqual([
@@ -663,10 +665,11 @@ describe('Snap Client', () => {
 						],
 						test: true,
 						siteId: '8uyt2m',
+						[NO_BEACON_PARAM]: true,
 					},
 				};
 
-				const recommendCacheKey = '{"profiles":[{"tag":"dress"}],"siteId":"8uyt2m","test":true}';
+				const recommendCacheKey = `{"profiles":[{"tag":"dress"}],"siteId":"8uyt2m","${NO_BEACON_PARAM}":true,"test":true}`;
 
 				expect(recommendRequesterSpy).toHaveBeenCalledTimes(2);
 				expect(recommendRequesterSpy.mock.calls).toEqual([