+++ packages/snap-preact-demo/templates/src/index.ts
@@ -3,45 +3,51 @@ import { globalStyles } from './styles';
 import deepmerge from 'deepmerge';
 import { combineMerge } from '../../snap/src/middleware/functions';
 import type { SnapTemplatesConfig } from '@searchspring/snap-preact';
-import { ClientConfig } from '@searchspring/snap-client';
+const siteId = 'atkzs2';
 
-const siteId = '8uyt2m';
+// const siteId = '8uyt2m';
 
-const clientConfig: ClientConfig = {
-	meta: {
-		origin: `https://${siteId}.a.searchspring.io`,
-	},
-	search: {
-		origin: `https://${siteId}.a.searchspring.io`,
-	},
-	autocomplete: {
-		requesters: {
-			suggest: {
-				origin: `https://${siteId}.a.searchspring.io`,
-			},
-			legacy: {
-				origin: `https://${siteId}.a.searchspring.io`,
-			},
-		},
-	},
-	finder: {
-		origin: `https://${siteId}.a.searchspring.io`,
-	},
-	recommend: {
-		origin: `https://${siteId}.a.searchspring.io`,
-	},
-	suggest: {
-		origin: `https://${siteId}.a.searchspring.io`,
-	},
-};
-
+// const clientConfig = {
+// 	meta: {
+// 		origin: `https://${siteId}.a.searchspring.io`,
+// 	},
+// 	search: {
+// 		origin: `https://${siteId}.a.searchspring.io`,
+// 	},
+// 	autocomplete: {
+// 		requesters: {
+// 			suggest: {
+// 				origin: `https://${siteId}.a.searchspring.io`,
+// 			},
+// 			legacy: {
+// 				origin: `https://${siteId}.a.searchspring.io`,
+// 			},
+// 		},
+// 	},
+// 	finder: {
+// 		origin: `https://${siteId}.a.searchspring.io`,
+// 	},
+// 	recommend: {
+// 		origin: `https://${siteId}.a.searchspring.io`,
+// 	},
+// 	suggest: {
+// 		origin: `https://${siteId}.a.searchspring.io`,
+// 	},
+// };
 let config: SnapTemplatesConfig = {
 	config: {
 		siteId,
 		language: 'en',
 		currency: 'usd',
 		platform: 'other',
-		client: clientConfig,
+		// client: clientConfig
+	},
+	plugins: {
+		common: {
+			addToCart: {
+				function: (data) => console.log('added to cart!', data),
+			},
+		},
 	},
 	components: {
 		result: {
@@ -63,7 +69,14 @@ let config: SnapTemplatesConfig = {
 			// },
 		},
 		style: globalStyles,
-		overrides: {},
+		overrides: {
+			// default: {
+			// 	'facet': {
+			// 		rangeInputs: true,
+			// 		rangeInputsPrefix: "$",
+			// 	}
+			// }
+		},
 	},
 	recommendation: {
 		email: {
@@ -89,6 +102,18 @@ let config: SnapTemplatesConfig = {
 				component: 'Search',
 			},
 		],
+		// settings: {
+		// 	filters: {
+		// 		fields: {
+		// 			'price': {
+		// 				filterFormatValue: '$%01.2f - $%01.2f'
+		// 			},
+		// 			'ss_category_hierarchy': {
+		// 				enabled: true,
+		// 			}
+		// 		}
+		// 	},
+		// },
 	},
 	autocomplete: {
 		targets: [
@@ -107,6 +132,7 @@ if (window.mergeSnapConfig) {
 new SnapTemplates(config);
 
 /*
+
 
 Overrides are taking priority over the theme layouts (responsive) specified within the Search component - but they shouldn't be.
 Look into: