+++ packages/snap-preact-demo/templates/src/universal.ts
@@ -3,14 +3,16 @@ import { polyfills } from '@searchspring/snap-preact';
 
 const promises = [];
 if (!('fetch' in window)) {
-	// @ts-ignore - types not important
-	promises.push(import('whatwg-fetch') as any);
+	promises.push(import('whatwg-fetch'));
 }
 if (!('Symbol' in window) || !('flatMap' in Array.prototype) || !('includes' in Array.prototype)) {
-	// @ts-ignore - types not important
-	promises.push(import('core-js/stable') as any);
+	promises.push(import('core-js/stable'));
 }
 promises.push(polyfills);
 Promise.all(promises).then(() => {
+	window.searchspring = window.searchspring || {};
+	window.searchspring.managed = true;
+	window.searchspring.build = 'universal';
+
 	import('./index');
 });