+++ packages/snap-preact/components/src/components/Organisms/FilterSummary/FilterSummary.stories.tsx
@@ -169,7 +169,7 @@ export default {
 const snapInstance = Snapify.search({
 	id: 'FilterSummary',
 	globals: {
-		siteId: '8uyt2m',
+		siteId: 'atkzs2',
 		filters: [
 			{
 				type: 'value',