+++ packages/snap-preact/components/tests/cypress/support/custom.js
@@ -1,0 +1,2 @@
+// used for ignoring certain runtime errors that may be unfixable
+export const ignoredErrors = ['Failed To Fetch'];