+++ packages/snap-preact/src/Templates/Stores/library/languages/es.ts
@@ -109,6 +109,9 @@ export const es: LangComponents = {
 		ctaButtonSuccessText: {
 			value: '¡Paquete agregado!',
 		},
+		ctaSubtotalTitle: {
+			value: (data) => `Subtotal para ${data.cartStore.count} artículos`,
+		},
 	},
 	mobileSidebar: {
 		openButtonText: {
@@ -169,6 +172,9 @@ export const es: LangComponents = {
 						(data?.facet as ValueFacet).values?.length ? (data?.facet as ValueFacet).values.length + ' opciones' : ''
 					}`,
 			},
+		},
+		submitRangeButton: {
+			value: 'Entregar',
 		},
 	},
 	select: {